Move Mixin classes out of main ``scoped_nodes`` file
